mapeando o item e pedido

+++ src/main/java/br/com/ithappens/model/ItemPedidoEstoque.java
@@ -3,11 +3,13 @@ package br.com.ithappens.model;
 import br.com.ithappens.mapper.typehandler.Status;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class ItemPedidoEstoque {
     private Long id;
     private Integer quantidadePedido;
-    private Integer valorDoPedido;
+    private BigDecimal valorDoPedido;
 
     //Relacionamento para a tabela de produtos
     private Produto produto;
