+++ src/main/java/br/com/ithappens/model/Produto.java
@@ -3,11 +3,13 @@ package br.com.ithappens.model;
 import br.com.ithappens.mapper.typehandler.Status;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class Produto {
   private Long id;
   private String descricao;
-  private Integer valorProduto;
+  private BigDecimal valorProduto;
 
   private Status status;
 }
